Change how PDFs are saved, add new options
Since Manual pages are loaded with a POST request, it's challenging to use Playwright to fetch them on its own. Therefore, the script used Axios to get the page's HTML, saved it as a .html file, loaded that .html file into Playwright, then PDF'd the loaded page.

This seems to various issues (see #12), so this commit switches to using `page.setContent`, then PDF-ing that.

.html files were originally saved because Playwright loaded them, which is no longer necessary. So, this commit no longer saves .html files by default. If users want that, they can use the two new command-line options, `--saveHTML` (which will save each page's HTML file) and `--ignoreSaveErrors`) which was requested in #12.